refactor: reorganize structure in the utils folder
