+++ src/utils/configs/index.js
@@ -1,0 +1,1 @@
+export * from './nav-items-config';
