+++ src/utils/constants/index.js
@@ -1,0 +1,2 @@
+export * from './movie-constants';
+export * from './routes-constants';
